tra_report_requirements.txt
# TraReport依赖包
torch>=1.9.0
torchvision>=0.10.0
omegaconf>=2.1.0
pytorch-lightning>=1.5.0
basicsr>=1.4.2
pillow>=8.0.0
tqdm>=4.60.0
opencv-python>=4.5.0
numpy>=1.21.0
einops>=0.3.0
